Move get_dir_size function to the payload.base.utils
This is payload specific code so it should be in the payload.
