+++ pyanaconda/modules/payload/base/initialization.py
@@ -23,7 +23,6 @@ from pyanaconda.core.constants import DD_ALL, DD_FIRMWARE, DD_RPMS
 from pyanaconda.modules.common.task import Task
 from pyanaconda.modules.payload.base.utils import create_root_dir, write_module_blacklist
 
-
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
