+++ pyanaconda/modules/payload/base/utils.py
@@ -18,9 +18,14 @@
 # Red Hat, Inc.
 #
 import os
+import stat
 
 from pyanaconda.core.util import mkdirChain
 from pyanaconda.flags import flags
+
+
+from pyanaconda.anaconda_loggers import get_module_logger
+log = get_module_logger(__name__)
 
 
 def create_root_dir(sysroot):
@@ -43,3 +48,44 @@ def write_module_blacklist(sysroot):
         f.write("# Module blacklists written by anaconda\n")
         for module in flags.cmdline["modprobe.blacklist"].split():
             f.write("blacklist %s\n" % module)
+
+
+def get_dir_size(directory):
+    """Get the size of a directory and all its subdirectories.
+
+    :param dir: the name of the directory to find the size of
+    :return: the size of the directory in kilobytes
+    """
+    def get_subdir_size(directory):
+        # returns size in bytes
+        try:
+            mydev = os.lstat(directory)[stat.ST_DEV]
+        except OSError as e:
+            log.debug("failed to stat %s: %s", directory, e)
+            return 0
+
+        try:
+            dirlist = os.listdir(directory)
+        except OSError as e:
+            log.debug("failed to listdir %s: %s", directory, e)
+            return 0
+
+        dsize = 0
+        for f in dirlist:
+            curpath = '%s/%s' % (directory, f)
+            try:
+                sinfo = os.lstat(curpath)
+            except OSError as e:
+                log.debug("failed to stat %s/%s: %s", directory, f, e)
+                continue
+
+            if stat.S_ISDIR(sinfo[stat.ST_MODE]):
+                if os.path.ismount(curpath):
+                    continue
+                if mydev == sinfo[stat.ST_DEV]:
+                    dsize += get_subdir_size(curpath)
+            elif stat.S_ISREG(sinfo[stat.ST_MODE]):
+                dsize += sinfo[stat.ST_SIZE]
+
+        return dsize
+    return get_subdir_size(directory) // 1024
